antigravity_backup/python_core/requirements.txt
langchain-openai
langchain-google-genai
chromadb
pypdf
reportlab
qrcode
python-dotenv
beautifulsoup4
requests
psycopg2-binary
